Abstract out QuestionForm as a reusable component
